emulate obscure ProTracker Dxx bug
Experimentation indicates that there’s a bug in ProTracker which causes
the Dxx command to jump one row too far if it was used on the same row
as an EEx command. Leaving this bug unemulated (like Milkytracker) can
cause problems with mods that rely on it. For example, Condom
Corruption (corruption.mod) by Travolta is missing a short section when
this bug is left unemulated.